mgr/cepahdm: move convert_tags_to_repo_digest to serve.py
